// backend/src/common/types.ts
export enum AuctionStatus {
  PENDING = 'pending',
  PUBLISHED = 'published',
  BIDDING = 'bidding',
  ENDED = 'ended',
  DEAL_CONFIRMED = 'deal_confirmed',
  PAYMENT_IN_PROGRESS = 'payment_in_progress',
  PAID = 'paid',
  EVICTION_IN_PROGRESS = 'eviction_in_progress',
  EVICTED = 'evicted',
  TAX_CALCULATED = 'tax_calculated',
  TRANSFERRED = 'transferred',
  COMPLETED = 'completed',
}